+++ src/ui/IWidgetContainer.cpp
@@ -92,7 +92,7 @@ void vui::IWidgetContainer::onMouseUp(Sender s, const MouseButtonEvent& e) {
 
 void vui::IWidgetContainer::onMouseMove(Sender s, const MouseMotionEvent& e) {
     if (!m_isEnabled) return;
-    if (isInBounds(e.x, e.y)) {
+    if (isInBounds((f32)e.x, (f32)e.y)) {
         if (!m_isMouseIn) {
             m_isMouseIn = true;
             MouseEnter(e);
